AR-UUID-To-Long modification des uuid par des long en tant qu'id #2

+++ src/main/java/dev/entites/dto/ReservationDto.java
@@ -26,7 +26,7 @@ public class ReservationDto {
 	private Long responsable_id;
 
 	@NotNull
-	private Integer vehicule_id;
+	private Long vehicule_id;
 
 	/**
 	 * Constructor
@@ -42,7 +42,7 @@ public class ReservationDto {
 	 * @param vehicule_id
 	 */
 	public ReservationDto(LocalDateTime dateDepart, LocalDateTime dateArrivee, String lieuDepart,
-			String lieuDestination, Integer dureeTrajet, Double distance, Long responsable_id, Integer vehicule_id) {
+			String lieuDestination, Integer dureeTrajet, Double distance, Long responsable_id, Long vehicule_id) {
 		super();
 		this.dateDepart = dateDepart;
 		this.dateArrivee = dateArrivee;
@@ -185,7 +185,7 @@ public class ReservationDto {
 	 *
 	 * @return the vehicule_id
 	 */
-	public Integer getVehicule_id() {
+	public Long getVehicule_id() {
 		return vehicule_id;
 	}
 
@@ -194,7 +194,7 @@ public class ReservationDto {
 	 *
 	 * @param vehicule_id the vehicule_id to set
 	 */
-	public void setVehicule_id(Integer vehicule_id) {
+	public void setVehicule_id(Long vehicule_id) {
 		this.vehicule_id = vehicule_id;
 	}
